requirements.txt
Flask==3.1.0
pymongo==4.10.1
python-dotenv==1.0.0
selenium==4.27.1
webdriver-manager==4.0.1
websocket-client==1.8.0